Update Detected Biometrics Changed for Android
It will invalidate the old private key if having any biometric enrollment after the moment we register in our app.
But we can not check with the biometricKeysExist function on Android, because it's just invalidated, not deleted.
So I suggest we should check the biometric changes for both Android and iOS by using the createSignature function. It will fail on both platforms with different error messages, but at least we can handle them in one place. Hope this help! Cheers!